Moved "reset security warnings" to advanced settings
