+++ app/src/main/res/xml/settings.xml
@@ -82,11 +82,6 @@
             android:title="@string/settings_hideTotalBalance"
             app:iconSpaceReserved="false" />
 
-        <Preference
-            android:key="resetGuardian"
-            android:title="@string/settings_resetSecurityWarnings"
-            app:iconSpaceReserved="false" />
-
     </PreferenceCategory>
